# docs/swagger.yaml
basePath: /
definitions:
...
  api.createTransactionRequest:
    properties:
      amount:
        type: string
      category_id:
        type: integer
      description:
        type: string
      transaction_date:
        type: string
      type:
        type: string
    required:
    - amount
    - transaction_date
    - type
    type: object
...
  api.transactionResponse:
    properties:
      amount:
        $ref: '#/definitions/pgtype.Numeric'
      category_id:
        type: integer
      created_at:
        $ref: '#/definitions/pgtype.Timestamptz'
      description:
        type: string
      id:
        type: integer
      transaction_date:
        $ref: '#/definitions/pgtype.Timestamptz'
      type:
        type: string
      user_id:
        type: integer
    type: object
...
  api.updateTransactionRequest:
    properties:
      amount:
        type: string
      category_id:
        type: integer
      description:
        type: string
      transaction_date:
        type: string
      type:
        type: string
    type: object
...
  big.Int:
    type: object
  pgtype.InfinityModifier:
    enum:
    - 1
    - 0
    - -1
    format: int32
    type: integer
    x-enum-varnames:
    - Infinity
    - Finite
    - NegativeInfinity
  pgtype.Numeric:
    properties:
      exp:
        format: int32
        type: integer
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      int:
        $ref: '#/definitions/big.Int'
      naN:
        type: boolean
      valid:
        type: boolean
    type: object
  pgtype.Timestamptz:
    properties:
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      time:
        type: string
      valid:
        type: boolean
    type: object
host: localhost:3000
info:
  contact: {}
  description: API untuk aplikasi pencatatan keuangan pribadi
  title: Catat API
  version: "1.0"
paths:
...
  /transaction:
    get:
      consumes:
      - application/json
      description: Mendapatkan daftar transaksi dengan pagination
      parameters:
      - description: 'Page number (min: 1)'
        in: query
        name: page_id
        required: true
        type: integer
      - description: 'Number of items per page (min: 5, max: 100)'
        in: query
        name: page_size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.transactionResponse'
            type: array
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List transactions
      tags:
      - Transactions
  /transactions:
    post:
      consumes:
      - application/json
      description: Membuat transaksi baru (pemasukan atau pengeluaran)
      parameters:
      - description: Create transaction request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.createTransactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.transactionResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Transaction already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create new transaction
      tags:
      - Transactions
  /transactions/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus transaksi berdasarkan ID
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Transaction deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Transaction not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete transaction
      tags:
      - Transactions
    get:
      consumes:
      - application/json
      description: Mendapatkan detail transaksi berdasarkan ID
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.transactionResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Transaction not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get transaction by ID
      tags:
      - Transactions
    put:
      consumes:
      - application/json
      description: Mengupdate transaksi berdasarkan ID
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update transaction request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.updateTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.transactionResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Transaction not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update transaction
      tags:
      - Transactions
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"